int64 tensor indexing: widen the (batch,head) slice offset to 64-bit so B*H*N*D may exceed 2^31; guard is now per-slice N*D<=INT32_MAX. Forward verified correct at base offsets above 2^31; backward matches the reference and is unchanged below it.

+++ csrc/kernels/backward/compute_dO3.cuh
@@ -22,7 +22,7 @@ __global__ void compute_dO3_kernel(
     scalar_t*       __restrict__ dO3,      // (BH, m, D) output in elem_type
     int D, int m
 ) {
-    const int bh = blockIdx.x;
+    const int64_t bh = blockIdx.x;
     const int tid = threadIdx.x;
     const int nthreads = blockDim.x;
